update storage description
