--- context/app/static/js/components/savedLists/LocalStorageDescription/LocalStorageDescription.tsx
@@ -1,35 +1,0 @@
-import React from 'react';
-import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import Stack from '@mui/material/Stack';
-import Typography from '@mui/material/Typography';
-
-import { useAppContext } from 'js/components/Contexts';
-import Description from 'js/shared-styles/sections/Description';
-import { InternalLink } from 'js/shared-styles/Links';
-
-function LocalStorageDescription() {
-  const { isAuthenticated } = useAppContext();
-
-  if (isAuthenticated) {
-    return <Description>Lists saved here are stored to your profile and accessible across devices.</Description>;
-  }
-
-  return (
-    <Description>
-      <Stack spacing={1}>
-        <Typography>
-          You must be logged in to access lists. <InternalLink href="/login">Log in</InternalLink> to access and manage
-          lists linked to your profile.
-        </Typography>
-        <Box>
-          <Button variant="contained" href="/login">
-            Log In
-          </Button>
-        </Box>
-      </Stack>
-    </Description>
-  );
-}
-
-export default LocalStorageDescription;
