--- context/app/static/js/components/savedLists/LocalStorageDescription/index.ts
@@ -1,3 +1,0 @@
-import LocalStorageDescription from './LocalStorageDescription';
-
-export default LocalStorageDescription;
