+++ context/app/static/js/components/savedLists/SavedListsDescription/index.ts
@@ -1,0 +1,3 @@
+import SavedListsDescription from './SavedListsDescription';
+
+export default SavedListsDescription;
